update lua globals

+++ backend/api/helpers/battlenet.js
@@ -220,7 +220,9 @@ module.exports = {
         char = new WoWChar({bnetID: summary.id, region, realm, realmSlug: summary.realm.slug, name})
       }
       else if (!char.bnetID !== summary.id) {
-        await WoWChar.findOneAndDelete({bnetID: summary.id})
+        if (summary.id) {
+          await WoWChar.findOneAndDelete({bnetID: summary.id})
+        }
         char = new WoWChar({bnetID: summary.id, region, realm, realmSlug: summary.realm.slug, name})
       }
 
@@ -286,7 +288,6 @@ module.exports = {
   },
 
   lookupCharacterStatus: async (region, realm, name) => {
-    console.log(region, realm, name)
     if (!realm) {
       return {}
     }
